split Dao, Database, Contract, and Mapper out
