+++ gto-support-db/src/androidTest/java/org/ccci/gto/android/common/db/QueryTest.java
@@ -4,7 +4,7 @@ import android.support.test.runner.AndroidJUnit4;
 import android.test.InstrumentationTestCase;
 import android.util.Pair;
 
-import org.ccci.gto.android.common.db.TestDao.TestContract.RootTable;
+import org.ccci.gto.android.common.db.Contract.RootTable;
 import org.junit.Test;
 import org.junit.runner.RunWith;
